Enable initial camera support.
Add "call_repeating" helper method which calls a function repeatedly, using a helper function to find out when to call it next.

Add a number of camera related calls into the setup which fetch images multiple times a day, and which enable/disable IR lights
on the cameras as sunup/sundown.