This is a large change.
Revamp config options for 'hosts'.

Revamped request handling for buttons. Revamped the config section for buttons.

Revamp how status interaction is handled to use status:// uris to reflect subsections of the current status.

Created a 'rules' engine to trigger actions based on changes to the status.
Created a 'rules' section in the config to configure the rules engine.