Remove all Log handling from the Status object.
Add Log extraction and packaging to the Log web resource class, since that's
the only place it's needed.

Switch all unittests to use test_base and the _create_status helper method.
This means there was only a single place to adjust all tests for the change
to the Status init signature.

Change the logging request url from /log_handler to /log.

Add cleanup code to both the Log and Status web resources so they don't
complain if a hanging web request is cut off remotely.