Add _IntervalHelper to rules engine.
Add _IntervalHelper to rules engine to handle Interval rules. Create matching
unittests.

All rules are not based on the _RuleHelper class, so all rules can not be
started and stopped. This finishes the prep work required to be able to
watch for changes to rules definitions and adjust the rules dynamically.