Add read-only sonos adapter. Made Sonos and SNMP adapters threaded.
Add a new basic Sonos adapter that displays status about Sonos. The SNMP
adapter now uses threading (like Sonos) to avoid blocking the main
thread. The Sonos adapter can be reworked to not need the threads, but
it's just not build that way yet.

The Sonos adapter is heavily based on the SoCo sample code.