Remove unused web_resources, remove Button 'action' support.
Remove the no-longer-used web resources 'Host' and 'Wake'.

Simplify 'Button' so that it no longer supports an attached action. We now
use rules that watch for changes in the 'pushed' value. This style of
interaction is more flexible, and removed the need for the adapter to decide
what happens when a button is pushed. It also simplifies out code.